+++ persistent-mongoDB/test/main.hs
@@ -104,8 +104,7 @@ EmptyEntity
 main :: IO ()
 main = do
   hspec $ afterAll dropDatabase $ do
-    xdescribe "This test is failing for Mongo by only embedding the first thing." $ do
-        RenameTest.specsWith (db' RenameTest.cleanDB)
+    RenameTest.specsWith (db' RenameTest.cleanDB)
     DataTypeTest.specsWith
         dbNoCleanup
         Nothing
@@ -137,13 +136,10 @@ main = do
         dbNoCleanup
         Nothing
     PersistentTest.specsWith (db' PersistentTest.cleanDB)
-    -- TODO: The upsert tests are currently failing. Find out why and fix
-    -- them.
-    xdescribe "UpsertTest is currently failing for Mongo due to differing behavior" $ do
-        UpsertTest.specsWith
-            (db' PersistentTest.cleanDB)
-            UpsertTest.AssumeNullIsZero
-            UpsertTest.UpsertGenerateNewKey
+    UpsertTest.specsWith
+        (db' PersistentTest.cleanDB)
+        UpsertTest.AssumeNullIsZero
+        UpsertTest.UpsertGenerateNewKey
     EmptyEntityTest.specsWith
         (db' EmptyEntityTest.cleanDB)
         Nothing